Service updates and improvements

+++ src/main/java/io/management/ua/certifications/service/CertificationService.java
@@ -6,14 +6,15 @@ import io.management.ua.amqp.models.messages.MessageModel;
 import io.management.ua.certifications.dto.CertificationDTO;
 import io.management.ua.certifications.entity.Certification;
 import io.management.ua.certifications.repository.CertificationRepository;
+import io.management.ua.exceptions.CertificationException;
 import io.management.ua.producers.CertificationResultProducer;
 import io.management.ua.producers.MessageProducer;
 import io.management.ua.utility.CodeGenerator;
 import io.management.ua.utility.TimeUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-import javax.transaction.Transactional;
 
 @Service
 @RequiredArgsConstructor
@@ -22,6 +23,7 @@ public class CertificationService {
     private final CertificationResultProducer certificationResultProducer;
     private final MessageProducer messageProducer;
 
+    @Transactional
     public void process(CertificationRequestModel certificationRequestModel) {
         String certificationCode = CodeGenerator.generateCode();
 
@@ -44,11 +46,11 @@ public class CertificationService {
         Certification certification = certificationRepository.findById(certificationDTO.getIdentifier())
                 .orElseThrow(() -> new RuntimeException(String.format("Certification was not found for identifier: %s", certificationDTO.getIdentifier())));
 
-        if (certification.getCode().equals(certificationDTO.getCode())) {
+        if (certification.getCode().equals(certificationDTO.getCode()) && certification.getIssuedAt().plusHours(1L).isAfter(TimeUtil.getCurrentTime())) {
             certificationRepository.deleteById(certificationDTO.getIdentifier());
             certificationResultProducer.produce(new CertificationResultModel(certificationDTO.getIdentifier(), true));
         } else {
-            certificationResultProducer.produce(new CertificationResultModel(certificationDTO.getIdentifier(), false));
+            throw new CertificationException(String.format("Code does not match, identifier: %s", certification.getIdentifier()));
         }
     }
